+++ src/personal_mnemonic_medium/exporters/anki/anki_card.py
@@ -1,75 +1,19 @@
 import copy
-import hashlib
 import os
 import re
 import urllib
 from pathlib import Path
-from typing import Any, List, Literal, Tuple
+from typing import Any, Callable, List, Literal, Tuple
 
 import genanki
-import misaka
 
 from personal_mnemonic_medium.exporters.anki.globals import CONFIG
+from personal_mnemonic_medium.exporters.markdown_to_html.html_compiler import (
+    compile_field,
+)
 from personal_mnemonic_medium.note_factories.note import Document
 from personal_mnemonic_medium.prompt_extractors.prompt import Prompt
-
-
-def strip_header(string: str) -> str:
-    """Strip first occurrence of a markdown level 1 header"""
-    return re.sub(r"^#.*\n", "", string)
-
-
-def field_to_html(field: Any) -> str:
-    # Math processing
-    """
-    Need to extract the math in brackets so that it doesn't get markdowned.
-    If math is separated with dollar sign it is converted to brackets.
-    """
-    if CONFIG["dollar"]:
-        for sep, (op, cl) in [("$$", (r"\\[", r"\\]")), ("$", (r"\\(", r"\\)"))]:
-            escaped_sep = sep.replace(r"$", r"\$")
-            # ignore escaped dollar signs when splitting the field
-            field = re.split(rf"(?<!\\){escaped_sep}", field)
-            # add op(en) and cl(osing) brackets to every second element of the list
-            field[1::2] = [op + e + cl for e in field[1::2]]
-            field = "".join(field)
-    else:
-        for bracket in ["(", ")", "[", "]"]:
-            field = field.replace(rf"\{bracket}", rf"\\{bracket}")
-            # backslashes, man.
-
-    for token in ["*", "/"]:
-        if token == "/":
-            replacement = "*"
-        elif token == "*":
-            replacement = "**"
-
-        pattern = f"\\{token}[^<>\\-\n]+\\{token}"
-
-        token_instances = re.findall(pattern, field)
-
-        for instance in token_instances:
-            field = field.replace(instance, replacement + instance[1:-1] + replacement)  # type: ignore
-
-    # Make sure every \n converts into a newline
-    field = field.replace("\n", "  \n")
-
-    return misaka.html(field, extensions=("fenced-code", "math"))  # type: ignore
-
-
-def compile_field(fieldtext: str) -> str:
-    """Turn source markdown into an HTML field suitable for Anki."""
-    fieldtext_sans_wiki = fieldtext.replace("[[", "<u>").replace("]]", "</u>")
-    fieldtext_sans_comments = re.sub(r"<!--.+-->", "", fieldtext_sans_wiki)
-
-    return field_to_html(fieldtext_sans_comments)
-
-
-def simple_hash(text: str) -> int:
-    """MD5 of text, mod 2^63. Probably not a great hash function."""
-    comp_hash = int(hashlib.sha256(text.encode("utf-8")).hexdigest(), 16) % 10**10
-
-    return comp_hash
+from personal_mnemonic_medium.utils.hasher import simple_hash
 
 
 class AnkiCard:
@@ -80,6 +24,8 @@ class AnkiCard:
         fields: List[str],
         source_prompt: Prompt,
         model_type: Literal["QA", "Cloze"],
+        html_compiler: Callable[[str], str],
+        url_generator: Callable[[str], str],
     ):
         self.markdown_fields = fields
         self.model_type = model_type
